Test move changes to another pc
