+++ backend/Tasque.Core.Common/DTO/PartialModels/TaskEntity/CosmosTaskFields.cs
@@ -5,11 +5,12 @@ using System.Text;
 using System.Threading.Tasks;
 using Tasque.Core.Common.Enums;
 
-namespace Tasque.Core.Common.DTO.PartialModels
+namespace Tasque.Core.Common.DTO.PartialModels.TaskEntity
 {
     public class CosmosTaskFields
     {
         public TaskFieldType FieldType { get; set; }
-        public string[]? FieldValues { get; set; }
+        public TaskCustomFields<string>[]? FieldValues { get; set; }
     }
 }
+
