+++ backend/Tasque.Core.Common/DTO/PartialModels/TaskEntity/TaskCosmosModel.cs
@@ -6,7 +6,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Tasque.Core.Common.StaticResources;
 
-namespace Tasque.Core.Common.DTO.PartialModels
+namespace Tasque.Core.Common.DTO.PartialModels.TaskEntity
 {
     public class TaskCosmosModel
     {
