+++ backend/Tasque.Core.Common/DTO/Project/NewProjectDto.cs
@@ -3,6 +3,7 @@
 public class NewProjectDto
 {
     public string? Name { get; set; }
+    public string Key { get; set; }
     public int AuthorId { get; set; }
     public int OrganizationId { get; set; }
 }
