+++ backend/Tasque.Core.Common/Entities/TaskPriority.cs
@@ -6,9 +6,9 @@ public class TaskPriority : BaseEntity
 {
     public TaskPriority()
     {
-        Tasks = new List<Task>();
+        //Tasks = new List<Task>();
     }
     public string Name { get; set; } = null!;
 
-    public virtual ICollection<Task> Tasks { get; set; }
+    //public virtual ICollection<Task> Tasks { get; set; }
 }
