+++ backend/Tasque.Core.Common/Entities/TaskState.cs
@@ -6,9 +6,9 @@ public class TaskState : BaseEntity
 {
     public TaskState()
     {
-        Tasks = new List<Task>();
+        //Tasks = new List<Task>();
     }
     public string Name { get; set; } = null!;
 
-    public virtual ICollection<Task> Tasks { get; set; }
+    //public virtual ICollection<Task> Tasks { get; set; }
 }
