+++ backend/Tasque.Core.Common/Entities/TaskType.cs
@@ -6,9 +6,9 @@ public class TaskType : BaseEntity
 {
     public TaskType()
     {
-        Tasks = new List<Task>();
+        //Tasks = new List<Task>();
     }
     public string Name { get; set; } = null!;
 
-    public virtual ICollection<Task> Tasks { get; set; }
+    //public virtual ICollection<Task> Tasks { get; set; }
 }
